Reorganize Folders and Refactorize Models
