--- src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Configuration/DatabaseShellConfigurations.cs
@@ -1,7 +1,0 @@
-namespace OrchardCore.Shells.Database.Configuration
-{
-    public class DatabaseShellConfigurations
-    {
-        public string Configurations { get; set; }
-    }
-}
