--- src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Configuration/DatabaseShellsConfigurations.cs
@@ -1,7 +1,0 @@
-namespace OrchardCore.Shells.Database.Configuration
-{
-    public class DatabaseShellsConfigurations
-    {
-        public string Configurations { get; set; }
-    }
-}
