--- src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Configuration/DatabaseShellsSettings.cs
@@ -1,7 +1,0 @@
-namespace OrchardCore.Shells.Database.Configuration
-{
-    public class DatabaseShellsSettings
-    {
-        public string Settings { get; set; }
-    }
-}
