+++ src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Models/DatabaseShellConfigurations.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json.Linq;
+
+namespace OrchardCore.Shells.Database.Models
+{
+    public class DatabaseShellConfigurations
+    {
+        public JObject ShellConfigurations { get; set; }
+    }
+}
