+++ src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Models/DatabaseShellsConfigurations.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json.Linq;
+
+namespace OrchardCore.Shells.Database.Models
+{
+    public class DatabaseShellsConfigurations
+    {
+        public JObject ShellsConfigurations { get; set; }
+    }
+}
