+++ src/OrchardCore/OrchardCore.Infrastructure/Shells.Database/Models/DatabaseShellsSettings.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json.Linq;
+
+namespace OrchardCore.Shells.Database.Models
+{
+    public class DatabaseShellsSettings
+    {
+        public JObject ShellsSettings { get; set; }
+    }
+}
